Switching from Arrays to database
In this update, I switch from storing data in arrays to storing them in a database.

I do this so that in the event that the server the code is running on reboots the code, the bot can get back up and running when it restarts because the database will still have all of the events that we need to keep track of stored. If the code reboots and I am storing that data locally in arrays instead, all of it would be lost upon rebooting, and the bot would be tweeting out things it has already said.